+++ bill/payment.go
@@ -3,7 +3,6 @@ package bill
 import (
 	"context"
 	"fmt"
-	"strconv"
 
 	"github.com/invopop/gobl/cal"
 	"github.com/invopop/gobl/cbc"
@@ -123,12 +122,20 @@ type Payment struct {
 	// to other documents and alternative parties involved in the order-to-delivery process.
 	Ordering *Ordering `json:"ordering,omitempty" jsonschema:"title=Ordering"`
 
-	// Summary of the taxes applied to the payment for tax regimes that require
-	// this information to be communicated.
-	Tax *tax.Total `json:"tax,omitempty" jsonschema:"title=Tax"`
+	// Total advances from any advances registered in the lines. Calculated automatically.
+	Advances *num.Amount `json:"advances,omitempty" jsonschema:"title=Advances,calculated=true"`
 	// Total amount to be paid in this payment, either positive or negative according to the
-	// line types and totals.
-	Total num.Amount `json:"total" jsonschema:"title=Total"`
+	// line types and totals. Calculated automatically.
+	Total num.Amount `json:"total" jsonschema:"title=Total,calculated=true"`
+	// Due reflects the amount that is still to be paid and will be calculated automatically
+	// based on the total and advance amounts, and may be require in some tax regimes or addons.
+	Due *num.Amount `json:"due,omitempty" jsonschema:"title=Due,calculated=true"`
+
+	// Summary of the taxes applied to the payment for tax regimes that require
+	// this information to be communicated. If payment lines contain `payable` amounts,
+	// these will be used to calculate the proportional amount of tax to apply automatically,
+	// otherwise the the taxes will be added together, assuming 100% is paid.
+	Tax *tax.Total `json:"tax,omitempty" jsonschema:"title=Tax,calculated=true"`
 
 	// Unstructured information that is relevant to the payment, such as correction or additional
 	// legal details.
@@ -285,27 +292,46 @@ func (pmt *Payment) calculate() error {
 		}
 	}
 
+	due := pmt.Currency.Def().Zero()
+	adv := pmt.Currency.Def().Zero()
 	for i, l := range pmt.Lines {
 		if l == nil {
 			continue
 		}
 		l.Index = i + 1
-		if err := l.calculate(pmt.Currency, pmt.ExchangeRates); err != nil {
-			return validation.Errors{
-				"lines": validation.Errors{
-					strconv.Itoa(i): err,
-				},
-			}
-		}
 
 		var lt *tax.Total
 		if l.Document != nil {
+			var er *currency.ExchangeRate
 			cur := l.Document.Currency
 			if cur == currency.CodeEmpty {
 				cur = pmt.Currency
+			} else {
+				// If the document has a currency, we need to ensure there is an exchange
+				// rate so any taxes can be converted correctly.
+				if er = currency.MatchExchangeRate(pmt.ExchangeRates, cur, pmt.Currency); er == nil {
+					return validation.Errors{
+						"exchange_rates": fmt.Errorf("%s to %s missing", cur, pmt.Currency),
+					}
+				}
 			}
-			l.Document.Calculate(cur, r.GetRoundingRule())
+			rr := r.GetRoundingRule()
+			l.Document.Calculate(cur, rr)
 			lt = l.Document.Tax.Clone()
+			lt.Exchange(er, rr)
+
+			// Perform extra calculations with the payable amount, if present.
+			if p := l.Document.Payable; p != nil {
+				// When calculating the taxes, determine if we need to rescale
+				// so that the taxes are proportional to the amount paid.
+				factor := l.Amount.Upscale(2).Divide(*p)
+				lt.Scale(factor, pmt.Currency, rr)
+				due = due.Add(*l.Document.Payable)
+				if l.Advances != nil {
+					due = due.Subtract(*l.Advances)
+				}
+				due = due.Subtract(l.Amount)
+			}
 
 			// Merge the line document taxes
 			if l.Document.Tax != nil {
@@ -318,18 +344,32 @@ func (pmt *Payment) calculate() error {
 		}
 
 		// Finally add the totals
-		a := l.Total
+		a := l.Amount
 		if total == nil {
 			total = &a
 		} else {
 			nt := total.Add(a)
 			total = &nt
 		}
+		if l.Advances != nil {
+			adv = adv.Add(*l.Advances)
+		}
 	}
-	pmt.Tax = tt
 	if total != nil {
 		pmt.Total = *total
 	}
+	if adv.IsZero() {
+		pmt.Advances = nil
+	} else {
+		pmt.Advances = &adv
+	}
+	if due.IsZero() {
+		pmt.Due = nil
+	} else {
+		pmt.Due = &due
+	}
+	pmt.Tax = tt
+
 	return nil
 }