+++ num/amount.go
@@ -150,15 +150,15 @@ func (a Amount) Split(x int) (Amount, Amount) {
 // Compare two amounts and return an integer value according to the
 // sign of the difference:
 //
-//	-1 if a <  a2
-//	 0 if a == a2
-//	 1 if a >  a2
-func (a Amount) Compare(a2 Amount) int {
-	a, a2 = rescaleAmountPair(a, a2)
-	if a.value < a2.value {
+//	-1 if a <  b
+//	 0 if a == b
+//	 1 if a >  b
+func (a Amount) Compare(b Amount) int {
+	a, b = rescaleAmountPair(a, b)
+	if a.value < b.value {
 		return -1
 	}
-	if a.value > a2.value {
+	if a.value > b.value {
 		return 1
 	}
 	return 0